fix: move glob edge case tests into abstract module (#1339)
